Move chart types into src/lib
So things in src/lib aren't digging around in src/components
